Extract React scaffold contract tests
